# river/internal/value: change representation of objects and maps