--- a/pkg/river/internal/stdlib/stdlib.go
+++ b/pkg/river/internal/stdlib/stdlib.go
@@ -4,12 +4,9 @@ package stdlib
 import (
 	"encoding/json"
 	"os"
-	"reflect"
 
 	"github.com/grafana/agent/pkg/river/internal/value"
 )
-
-var goAny = reflect.TypeOf((*interface{})(nil)).Elem()
 
 // Functions returns the list of stdlib functions by name. The interface{}
 // value is always a River-compatible function value, where functions have at
@@ -53,39 +50,14 @@ var Functions = map[string]interface{}{
 			return args[0], nil
 		}
 
-		// If the imcoming Go slices have the same type, we can have our resulting
-		// slice use the same type. This will allow decoding to use the direct
-		// assignment optimization.
-		//
-		// However, if the types don't match, then we're forced to fall back to
-		// returning []interface{}.
-		//
-		// TODO(rfratto): This could fall back to checking the elements if the
-		// array/slice types don't match. It would be slower, but the direct
-		// assignment optimization probably justifies it.
-		useType := args[0].Reflect().Type()
-		for i := 1; i < len(args); i++ {
-			if args[i].Reflect().Type() != useType {
-				useType = reflect.SliceOf(goAny)
-				break
-			}
-		}
-
-		// Build out the final array.
-		raw := reflect.MakeSlice(useType, finalSize, finalSize)
-
-		var argNum int
+		raw := make([]value.Value, 0, finalSize)
 		for _, arg := range args {
 			for i := 0; i < arg.Len(); i++ {
-				elem := arg.Index(i)
-				if elem.Type() != value.TypeNull {
-					raw.Index(argNum).Set(elem.Reflect())
-				}
-				argNum++
+				raw = append(raw, arg.Index(i))
 			}
 		}
 
-		return value.Encode(raw.Interface()), nil
+		return value.Array(raw...), nil
 	}),
 
 	"json_decode": func(in string) (interface{}, error) {