This commit changes the representation of objects and maps from `[]interface{}` and `map[string]interface{}` to use `[]value.Value` and `map[string]value.Value` respectively.

This fixes #2362, an issue where interface-based capsule values where losing type information and being coerced to a type based on the actual value backing the interface.

This also greatly improves the performance of creating arrays and objects, which no longer need to be inidividually copied in a loop:

```
name                   old time/op    new time/op    delta
Object/Non-capsule-10    1.84µs ± 8%    0.00µs ± 0%   -99.89%  (p=0.000 n=19+19)
Object/Capsule-10        1.82µs ± 4%    0.00µs ± 0%   -99.89%  (p=0.000 n=20+19)
Array/Non-capsule-10      389ns ± 2%      19ns ± 2%   -95.20%  (p=0.000 n=19+18)
Array/Capsule-10          389ns ± 1%      19ns ± 0%   -95.19%  (p=0.000 n=20+20)

name                   old alloc/op   new alloc/op   delta
Object/Non-capsule-10    1.88kB ± 0%    0.00kB       -100.00%  (p=0.000 n=20+20)
Object/Capsule-10        1.88kB ± 0%    0.00kB       -100.00%  (p=0.000 n=20+20)
Array/Non-capsule-10       344B ± 0%       24B ± 0%   -93.02%  (p=0.000 n=20+20)
Array/Capsule-10           344B ± 0%       24B ± 0%   -93.02%  (p=0.000 n=20+20)

name                   old allocs/op  new allocs/op  delta
Object/Non-capsule-10      42.0 ± 0%       0.0       -100.00%  (p=0.000 n=20+20)
Object/Capsule-10          42.0 ± 0%       0.0       -100.00%  (p=0.000 n=20+20)
Array/Non-capsule-10       2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=20+20)
Array/Capsule-10           2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=20+20)
```

As part of this change, the logic used for the concat function has been greatly simplified. This has seemingly no impact on performance, as the benchmark results below show:

```
name       old time/op    new time/op    delta
Concat-10    1.10µs ± 4%    1.09µs ± 1%   ~     (p=0.735 n=18+19)

name       old alloc/op   new alloc/op   delta
Concat-10    1.30kB ± 0%    1.30kB ± 0%   ~     (all equal)

name       old allocs/op  new allocs/op  delta
Concat-10      20.0 ± 0%      20.0 ± 0%   ~     (all equal)
```

I didn't update the CHANGELOG because #2362 only manifested for the `otelcol` components, which haven't been released yet.