@rfratto 2022-10-14
> This is way simpler and cleaner. A little surprised there are no gotchas but looks great!

Yeah, I'm a little nervous that I'm missing something and that something is going to break after this is merged, but we'll see what happens 🤷‍♂️